pyproject.toml
[tool.poetry]
name = "musichubdrf"
version = "0.1.0"
description = ""
authors = ["Damian Bednarek <damian.bednarekk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.10"
Django = "~4.1"
django-configurations = "~2.4"
djangorestframework = "~3.14"
Markdown = "~3.4"
django-filter = "~22.1"
django-rest-authemail = "~2.1"
factory-boy = "~3.2"
coverage = "~7.0"
psycopg2-binary = "~2.9"
Pillow = "~9.4"
mutagen = "~1.46"
django-storages = "~1.13"
boto3 = "~1.26"
gunicorn = "~20.1"
requests = "~2.28"
pydantic = "1.10"
drf-social-oauth2 = "~1.2"
drf-yasg = "~1.21"
django-extensions = "~3.2"
djangorestframework-simplejwt = "^5.2.2"
pytest = "^7.2.2"
pytest-django = "^4.5.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
